Authorization

+++ SecureDataApp/Controllers/ContactsController.cs
@@ -2,41 +2,56 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Formatters;
+using SecureDataApp.Authorization;
 using SecureDataApp.Data;
 using SecureDataApp.Data.Repository;
 using SecureDataApp.Models;
 
 namespace SecureDataApp.Controllers
 {
+    [Authorize]
     public class ContactsController : Controller
     {
 
         private readonly IContactRepository _contactRepository;
+        private readonly UserManager<IdentityUser> _userManager;
+        private readonly IAuthorizationService _authorizationService;
 
-        public ContactsController(IContactRepository contactRepository)
+        public ContactsController(IContactRepository contactRepository, 
+                                    UserManager<IdentityUser> userManager,
+                                    IAuthorizationService authorizationService)
         {
             _contactRepository = contactRepository;
+            _userManager = userManager;
+            _authorizationService = authorizationService;
+
         }
 
-        public IActionResult Index()
+        public async Task<IActionResult> Index()
         {
-            IEnumerable<Contact> _contacts = _contactRepository.ListAllContacts();
-            return View(_contacts);
+            var user = await _userManager.GetUserAsync(User);
+            IEnumerable<Contact> contacts = _contactRepository.ListAllContacts(user);
+            return View(contacts);
         }
 
         [HttpGet("contacts/details/{contactId}")]
-        public IActionResult Details(int contactId)
+        public async Task<IActionResult> Details(int contactId)
         {
-            Contact _contact = _contactRepository.ListContact(contactId);
-            if (_contact != null)
-            {
-                return View(_contact);
-            }
-            else
+            Contact contact = _contactRepository.ListContact(contactId);
+            if (contact == null)
             {
                 return NotFound();
             }
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, contact, ContactOperations.Read);
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
+            return View(contact);
         }
 
         [HttpGet]
@@ -47,53 +62,86 @@ namespace SecureDataApp.Controllers
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public IActionResult Add(Contact contact)
+        public async Task<IActionResult> Add(Contact contact)
         {
+            if (!ModelState.IsValid)
+            {
+                return ValidationProblem();
+            }
+            var user = await _userManager.GetUserAsync(User);
+            contact.UserId = user.Id;
             _contactRepository.AddContact(contact);
             return RedirectToAction("Index");
         }
 
         [HttpGet("contacts/edit/{contactId}")]
-        public IActionResult Edit(int contactId)
+        public async Task<IActionResult> Edit(int contactId)
         {
-            Contact _contact = _contactRepository.ListContact(contactId);
-            if (_contact != null)
+            Contact contact = _contactRepository.ListContact(contactId);
+            var authorizationResult =
+                await _authorizationService.AuthorizeAsync(User, contact, ContactOperations.Update);
+            if (!authorizationResult.Succeeded)
             {
-                return View(_contact);
+                return Forbid();
             }
-            else
-            {
-                return NotFound();
-            }
+
+            return View(contact);
+            
         }
 
-        [HttpPost]
+        [HttpPost("contacts/edit/{contactId}"), ActionName("Edit")]
         [ValidateAntiForgeryToken]
-        public IActionResult Update(Contact contact)
+        public async Task<IActionResult> Update(int contactId ,Contact contact)
         {
+            var query = _contactRepository.FindNoTrackingContact(contactId);
+            if (query.ContactId != contact.ContactId)
+            {
+                return BadRequest();
+            }
+            var authorizationResult =
+                await _authorizationService.AuthorizeAsync(HttpContext.User, contact, ContactOperations.Update);
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
+            if (!ModelState.IsValid)
+            {
+                return ValidationProblem();
+            }
             _contactRepository.EditContact(contact);
             return RedirectToAction("Index");
         }
 
+        
         [HttpGet("contacts/delete/{contactId}")]
-        public IActionResult Delete(int contactId)
+        public async Task<IActionResult> Delete(int contactId)
         {
-            Contact _contact = _contactRepository.ListContact(contactId);
-            if (_contact != null)
+            Contact contact = _contactRepository.ListContact(contactId);
+            var authorizationResult =
+                await _authorizationService.AuthorizeAsync(User, contact, ContactOperations.Delete);
+            if (!authorizationResult.Succeeded)
             {
-                return View(_contact);
+                return Forbid();
             }
-            else
-            {
-                return NotFound();
-            }
+            return View(contact);
         }
 
-        [HttpPost, ActionName("Delete")]
-        public IActionResult Destroy(Contact contact)
+        
+        [HttpPost("contacts/delete/{contactId}"), ActionName("Delete")]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Destroy(int contactId)
         {
+            Contact contact = _contactRepository.ListContact(contactId);
+            var authorizationResult =
+                await _authorizationService.AuthorizeAsync(HttpContext.User, contact, ContactOperations.Delete);
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
+            
             _contactRepository.DeleteContact(contact);
             return RedirectToAction("Index");
+            
         }
 
     }
